Create page for collateral tvl deposit
